Delta Hedging Simulator
Simulatior of a delta heding strategy for a portfolio constructed by a short position in a European call option and a long postion in stock. The postion on stock is defined by the hedging parameter. The cost and profitability of the trading strategy are computed for two different hedging adjustment intervals. The effect of mismatching volatility between underlying and option is investigated.